requirements-dev.txt
-r requirements-dep.txt
django-debug-toolbar==1.10.1
factory-boy==2.8.1
coverage==4.4.1
coveralls==1.2.0
